requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.2
Flask-Migrate==4.1.0
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.49
Werkzeug==3.1.8